[SELC-7291] feat: add geographic taxonomy management to DashboardOverview and PartyDetail components
