+++ src/services/__mocks__/partyService.ts
@@ -660,10 +660,7 @@ export const mockedParties: Array<Party> = [
     origin: 'IPA',
     institutionType: 'PA',
     recipientCode: 'CGDAS23D',
-    geographicTaxonomies: [
-      { code: '5555678', desc: 'Comune di Milano' },
-      { code: '8853441', desc: 'Comune di Roma' },
-    ],
+    geographicTaxonomies: [],
     vatNumber: '111122211111',
     products: [
       {
